% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{add_program_score}
\alias{add_program_score}
\title{Add gene program scores}
\usage{
add_program_score(dataset, features, org = "human", nbin = 20, ctrl = 10, name)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{features}{An string vector - a gene list of expression programs.}

\item{org}{A string - name of organism. Currently only "human" or "mouse" are accepted.}

\item{nbin}{An integer - number of bins of aggregate expression levels for all analyzed features}

\item{ctrl}{An integer - number of control features selected from the same bin per analyzed feature}

\item{name}{A string - name of the expression program}
}
\value{
A Seurat object.
}
\description{
Add gene program scores
}
