% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{analyze_merged}
\alias{analyze_merged}
\title{Streamlined analysis of Seurat object after integration}
\usage{
analyze_merged(dataset, group.levels, verbose = T, npcs = 50,
  reduction = "umap", dims = 1:20, nnei = 30, min.dist = 0.3,
  spread = 1, n.epochs = 500, k.param = 20, resolution = 0.8)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{group.levels}{A string vector - factor levels of biological samples.}

\item{verbose}{A logical value -}

\item{npcs}{An integer -}

\item{reduction}{A string -}

\item{dims}{An integer vector -}

\item{nnei}{An integer -}

\item{min.dist}{A double -}

\item{spread}{A double -}

\item{n.epochs}{An integer -}

\item{k.param}{An integer -}

\item{resolution}{A double -}
}
\value{
A Seurat object.
}
\description{
Streamlined analysis of Seurat object after integration
}
