% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{are_colors}
\alias{are_colors}
\title{Test if a vector of strings are valid colors}
\usage{
are_colors(x)
}
\arguments{
\item{x}{A string vector.}
}
\value{
A named vector of logicals.
}
\description{
Test if a vector of strings are valid colors
}
