% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_symbol}
\alias{convert_symbol}
\title{Conversion between mouse and human gene symbols}
\usage{
convert_symbol(gene_list, hs_to_mm = TRUE)
}
\arguments{
\item{gene_list}{A string vector - a list of genes to convert}

\item{hs_to_mm}{A logical value - if the conversion is from human to mouse. Default value is \code{TRUE}.}
}
\value{
A string vector.
}
\description{
Conversion between mouse and human gene symbols
}
