% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_norm_10X}
\alias{filter_norm_10X}
\title{Filter and normalize Seurat object}
\usage{
filter_norm_10X(dataset, nfeature = 500, mito = 10, nfeatures = 2000)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{nfeature}{An integer - include cells with at least this many features.}

\item{mito}{An integer - include cells with at most this many percentage of mitochondrial RNA molecules.}

\item{nfeatures}{An integer - number of features to select as top variable features.}
}
\value{
A Seurat object.
}
\description{
Filter and normalize Seurat object
}
