% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_scdata}
\alias{filter_scdata}
\title{Filter Seurat object}
\usage{
filter_scdata(
  dataset,
  range_nFeature = c(500, Inf),
  range_nCount = c(500, Inf),
  range_mt = c(-Inf, 10)
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{range_nFeature}{An integer vector - the range of nFeature_RNA for filtering. Default value is \code{c(500, Inf)}.}

\item{range_nCount}{An integer vector - the range of nCount_RNA for filtering. Default value is \code{c(500, Inf)}.}

\item{range_mt}{An numeric vector - the range of percent.mt for filtering. Default value is \code{c(-Inf, 10)}.}
}
\value{
A Seurat object.
}
\description{
Filter Seurat object
}
