% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_scdata_list}
\alias{filter_scdata_list}
\title{Filter Seurat object list}
\usage{
filter_scdata_list(data_list, ...)
}
\arguments{
\item{data_list}{A list of Seurat objects.}

\item{...}{Arguments passed to \code{filter_scdata}.}
}
\value{
A list of Seurat objects.
}
\description{
Filter Seurat object list
}
