% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_diff_genes}
\alias{find_diff_genes}
\title{Find differential expressed genes between two groups in each cluster}
\usage{
find_diff_genes(dataset, clusters, comparison, ...)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{clusters}{A character vector - clusters to investigate.}

\item{comparison}{A character vector of length 3 - the first element is the variable name, the second one is the level of the baseline group, 
the third one is the level to be compared with the baseline.

For example, \code{c("condition","untreated","treated")} will generate the results of \code{"treated"} group vs \code{"untreated"} group in the \code{"condition"} variable.}

\item{...}{Additional parameters to be passed to \code{Seurat::FindMarkes}.}
}
\value{
A dataframe.
}
\description{
Find differential expressed genes between two groups in each cluster
}
