% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gene_in_data}
\alias{gene_in_data}
\title{Test whether a gene waas in the dataset}
\usage{
gene_in_data(dataset, gene)
}
\arguments{
\item{dataset}{A Seurat object -}

\item{gene}{A string -}
}
\value{
A double
}
\description{
Test whether a gene waas in the dataset
}
