% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_colors}
\alias{get_colors}
\title{Get a vector of colors.}
\usage{
get_colors(v, pal = "Paired")
}
\arguments{
\item{v}{An integer index vector.}

\item{pal}{A string value describing the name of palette in \code{RColorBrewer}.}
}
\value{
A vector of colors.
}
\description{
Get a vector of colors.
}
\examples{
get_colors(1:5, "Set1")
get_colors(c(1:2,4,6), "Paired")
}
