% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_gene_data}
\alias{get_gene_data}
\title{Obtain the expression data for genes}
\usage{
get_gene_data(dataset, genes)
}
\arguments{
\item{dataset}{A Seurat object -}

\item{genes}{A string vector -}
}
\value{
A tibble
}
\description{
Obtain the expression data for genes
}
