% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_measure_data}
\alias{get_measure_data}
\title{Obtain the expression data for genes}
\usage{
get_measure_data(dataset, measures, return_df = TRUE)
}
\arguments{
\item{dataset}{A Seurat object -}

\item{measures}{A string vector -}

\item{return_df}{A logical - if \code{TRUE}, return the dataframe; if \code{FALSE}, return a list of the dataframe and valid measure names. Default value is \code{TRUE}.}
}
\value{
A tibble
}
\description{
Obtain the expression data for genes
}
