% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Obtain the meta data for features}
\usage{
get_meta_data(dataset, measures)
}
\arguments{
\item{dataset}{A Seurat object -}

\item{measures}{A string vector -}
}
\value{
A tibble
}
\description{
Obtain the meta data for features
}
