% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_palette}
\alias{get_palette}
\title{Get a vector of colors for cluster coding}
\usage{
get_palette(ncolor, palette = c("Paired", "Set2", "Set1"))
}
\arguments{
\item{ncolor}{Number of colors to be generated}

\item{palette}{A string vector describing the name of palette in \code{RColorBrewer}.}
}
\value{
A vector of colors.
}
\description{
Get a vector of colors for cluster coding
}
\examples{
get_palette(12, c("Set2", "Set1"))

}
