% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_spectrum}
\alias{get_spectrum}
\title{Get a vector of colors for sample coding}
\usage{
get_spectrum(n, palette = "Set2")
}
\arguments{
\item{n}{Number of colors to be generated}
}
\value{
A vector of colors
}
\description{
Get a vector of colors for sample coding
}
