% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_top_genes}
\alias{get_top_genes}
\title{Get top genes of each cluster}
\usage{
get_top_genes(dataset, markers, n)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{markers}{A tibble - from FindMarkers function.}

\item{n}{An integer - number of top genes to choose.}
}
\value{
A string vector of genes.
}
\description{
Get top genes of each cluster
}
