% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{load_10X_mito_cc}
\alias{load_10X_mito_cc}
\title{Get a vector of colors}
\usage{
load_10X_mito_cc(dir = NULL, raw_data = NULL, gcol = 2, proj_name,
  min_cells = 5, org = "human", cc = T)
}
\arguments{
\item{dir}{A string - directory path generated by 10xGenomics pipeline.}

\item{raw_data}{A string - file path of single cell RNA-Seq count matrix.}

\item{gcol}{An integer - the column index of feature names in the 10xGenomics folder.}

\item{proj_name}{A string - name of the biological sample.}

\item{min_cells}{An integer - include features detected in at least this many cells.}

\item{org}{A string - name of organism. Currently only "human" or "mouse" are accepted.}

\item{cc}{A logcial value - if cell cycle scores are calculated}
}
\value{
A Seurat of object.
}
\description{
Get a vector of colors
}
