% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_all_cluster_go}
\alias{plot_all_cluster_go}
\title{plot the GO enrichment analysis of all clusters of a dataset}
\usage{
plot_all_cluster_go(markers, ...)
}
\arguments{
\item{markers}{A tibble -}

\item{...}{Additional arguments to be passed to the function \code{\link{plot_cluster_go}}.}
}
\value{
A plot.
}
\description{
plot the GO enrichment analysis of all clusters of a dataset
}
