% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Single cell visualization - colored by cluster}
\usage{
plot_cluster(dataset, label = T, levels = NULL, self_set_color = F,
  self_colors, palette = c("Set2", "Paired"))
}
\arguments{
\item{dataset}{A Seurat object.}

\item{label}{A logical value -}

\item{levels}{A string vector -}

\item{self_set_color}{A logical value -}

\item{self_colors}{A string vector -}

\item{palette}{A string vector -}

\item{reduction}{A string -}
}
\value{
A plot.
}
\description{
Single cell visualization - colored by cluster
}
