% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cluster_go}
\alias{plot_cluster_go}
\title{plot the GO enrichment analysis of a cluster}
\usage{
plot_cluster_go(markers, cluster_name, topn = 100, org, ...)
}
\arguments{
\item{markers}{A tibble -}

\item{cluster_name}{A string -}

\item{topn}{An integer -}

\item{org}{A string -}

\item{...}{Additional arguments to be passed to the function \code{\link[clusterProfiler]{enrichGO}}}
}
\value{
A plot.
}
\description{
plot the GO enrichment analysis of a cluster
}
