% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_features}
\alias{plot_features}
\title{Single cell visualization - gene expression levels}
\usage{
plot_features(dataset, features, ncol)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{features}{A string vector -}

\item{ncol}{An integer -}
}
\value{
A plot.
}
\description{
Single cell visualization - gene expression levels
}
