% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_GSEA}
\alias{plot_GSEA}
\title{plot the results of GSEA}
\usage{
plot_GSEA(gsea_res, pattern = "HALLMARK_", p_cutoff = 0.05)
}
\arguments{
\item{gsea_res}{A tibble -}

\item{pattern}{A string -}

\item{p_cutoff}{A double -}
}
\value{
A plot.
}
\description{
plot the results of GSEA
}
