% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_GSVA}
\alias{plot_GSVA}
\title{plot the GSVA scores of given gene signatures}
\usage{
plot_GSVA(dataset, gene_list, pattern = "HALLMARK_", ...)
}
\arguments{
\item{dataset}{A Seurat object}

\item{gene_list}{A list -}

\item{pattern}{A string -}

\item{...}{Additional arguments to be passed to the function \code{\link[GSVA]{gsva}}}
}
\value{
A plot.
}
\description{
plot the GSVA scores of given gene signatures
}
