% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap of single cell dataset}
\usage{
plot_heatmap(
  dataset,
  markers,
  sort_var = c("seurat_clusters"),
  n = 8,
  anno_var,
  anno_colors,
  hm_limit = c(-2, 0, 2),
  hm_colors = c("#4575b4", "white", "#d73027")
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{markers}{A tibble - the dataframe generated by \code{FindAllMarkers} function in \code{Seurat};

Or a character vector - the names of genes can be directly specified.}

\item{sort_var}{A string vector - the variables to be sorted in the heatmap, should exist in the \code{metadata} of the \code{dataset}.}

\item{n}{An integer - number of genes to be plotted for each \code{seurat_cluster}. This parameter will not be used if the \code{markers} are directly specified.}

\item{anno_var}{A string vector - the variables in the annotation, should exist in the \code{metadata} of the \code{dataset}.}

\item{anno_colors}{A list - the color specification of annotation bars. The length should be the same as \code{anno_var}, with each list element corresponding to each variable. 

For a numeric variable, the element can be the name of a sequential or divergent palette in \code{RColorBrewer}. Please use \code{brewer.pal.info} for more information. 
It can also be a string vector with three color values, corresponding to the \code{min}, \code{median}, and \code{max} values of the numeric variable. 

For a categorical variable, the element can be the names of palettes in \code{RColorBrewer}. Multiple palettes can be specified, in case of many levels to be colored. 
It can also be a string vector of colors, corresponding to the levels in the categorical variable. Hence, the number of colors specified should be at least the number of levels (unique values) in the categorical variable.}

\item{hm_limit}{A numeric vector - three numeric values that dictate the lowest limit, midpoint, and highest limit of the color gradient. These three values correspond to the colors specified in the \code{hm_colors}.}

\item{hm_colors}{A string vector - three colors corresponding to the lowest limit, midpoint, and highest limit specified in \code{hm_colors}.}
}
\value{
A plot.
}
\description{
Plot the heatmap of single cell dataset
}
