% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_measure}
\alias{plot_measure}
\title{Boxplot of gene program scores}
\usage{
plot_measure(dataset, measure, plot_type, group_levels, cluster_levels)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{measure}{A string -}

\item{plot_type}{A string -}

\item{group_levels}{A string vector -}

\item{cluster_levels}{A string vector -}
}
\value{
A plot.
}
\description{
Boxplot of gene program scores
}
