% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_measure_cluster}
\alias{plot_measure_cluster}
\title{Single cell visualization - gene expression levels}
\usage{
plot_measure_cluster(dataset, features, plot_type = "none", meta = FALSE)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{features}{A string vector -}

\item{plot_type}{A string -}

\item{meta}{A logical value -}
}
\value{
A plot.
}
\description{
Single cell visualization - gene expression levels
}
