% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_measure_dim}
\alias{plot_measure_dim}
\title{Single cell visualization - gene expression levels}
\usage{
plot_measure_dim(dataset, measures, split_by = NA, point.size = 1)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{measures}{A character vector - the names of genes or measures to be plotted.}

\item{split_by}{A string - the splitting variable in the metadata.}

\item{point.size}{A numeric value - the size of points.}
}
\value{
A plot.
}
\description{
Single cell visualization - gene expression levels
}
