% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_merge}
\alias{plot_merge}
\title{Single cell visualization - merged and colored by biological sample}
\usage{
plot_merge(dataset, colors = NULL, ...)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{colors}{A string vector - the colors used for samples. Default value is \code{NULL}.}

\item{...}{Arguments passed to \code{DimPlot}.}
}
\value{
A plot.
}
\description{
Single cell visualization - merged and colored by biological sample
}
