% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_qc}
\alias{plot_qc}
\title{Plot the QC metrics of an individual biological sample}
\usage{
plot_qc(data_list, metrics, plot_type = "combined", palette = "Set2")
}
\arguments{
\item{data_list}{A list of Seurat objects}

\item{metrics}{A string - the name of the QC metrics}

\item{plot_type}{A string - the type of the plot. Default value is \code{"combined"}.}

\item{palette}{A string - the \code{RColorBrewer} palette to be used. Default value is \code{"Set2"}.}
}
\value{
A plot.
}
\description{
Plot the QC metrics of an individual biological sample
}
