% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_scdata}
\alias{plot_scdata}
\title{Single cell visualization - merged and colored by biological sample}
\usage{
plot_scdata(
  dataset,
  color_by = "seurat_clusters",
  split_by = NA,
  pal_setup = "Set2"
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{color_by}{A string - by which metadata variable the colors will be applied.}

\item{split_by}{A string - the splitting variable in the metadata.}

\item{pal_setup}{A dataframe with 2 columns - the \code{RColorBrewer} palette setup to be used;

Or a character vector of \code{RColorBrewer} palettes. Multiple palettes can be specified, in case of many levels to be colored; 

Or a character vector of colors.}
}
\value{
A plot.
}
\description{
Single cell visualization - merged and colored by biological sample
}
