% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_split}
\alias{plot_split}
\title{Single cell visualization - split by biological sample, colored by cluster}
\usage{
plot_split(dataset, reduction = "umap", split.by = "group",
  colors = c("#92c5de", "#d6604d"), legend.title = "Cluster",
  labels = levels(dataset$seurat_clusters))
}
\arguments{
\item{dataset}{A Seurat object.}

\item{reduction}{A string -}

\item{split.by}{A string -}

\item{colors}{A string vector -}

\item{legend.title}{A string -}

\item{labels}{A string vector -}
}
\value{
A plot.
}
\description{
Single cell visualization - split by biological sample, colored by cluster
}
