% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{refactor_seurat}
\alias{refactor_seurat}
\title{Refactor grouping variables in Seurat object}
\usage{
refactor_seurat(dataset, metadata = NULL)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{metadata}{A dataframe - should be the same as the one used during loading. Default value is \code{NULL}.}
}
\value{
A Seurat object.
}
\description{
Refactor grouping variables in Seurat object
}
