% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{rename_cluster}
\alias{rename_cluster}
\title{Rename clusters in Seurat object}
\usage{
rename_cluster(dataset, labels)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{labels}{An string vector - new names of each cluster.}
}
\value{
A Seurat object.
}
\description{
Rename clusters in Seurat object
}
