% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_colors}
\alias{set_colors}
\title{Setup the colors for plotting}
\usage{
set_colors(pal, n)
}
\arguments{
\item{pal}{A string or a string vector.}

\item{n}{An integer.}
}
\value{
A string vector of colors.
}
\description{
Setup the colors for plotting
}
