% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{test_GSEA}
\alias{test_GSEA}
\title{GSEA of differential gene expression in each cluster}
\usage{
test_GSEA(diff, clusters = NULL, pathway)
}
\arguments{
\item{diff}{A dataframe - the dataframe generated by \code{Seurat::FindMarkers}.}

\item{clusters}{A character vector - clusters to investigate.}

\item{pathway}{A list of character vectors - gene lists for GSEA. The names of list elements should be the pathway names.}
}
\value{
A dataframe.
}
\description{
GSEA of differential gene expression in each cluster
}
