% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vlookup}
\alias{vlookup}
\title{Look up corresponding values from a lookup table}
\usage{
vlookup(list, table, col_1, col_2)
}
\arguments{
\item{list}{A string vector - keys}

\item{table}{A tibble - the lookup table}

\item{col_1}{An integer - column index of the keys}

\item{col_2}{An integer - column index of the values}
}
\value{
A string vector.
}
\description{
Look up corresponding values from a lookup table
}
