% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_scdata}
\alias{filter_scdata}
\title{Filter a list of Seurat objects}
\usage{
filter_scdata(data_list, ...)
}
\arguments{
\item{data_list}{A list of Seurat objects}

\item{...}{Additional arguments to be passed to the function \code{subset}.}
}
\value{
A list of Seurat objects
}
\description{
Filter a list of Seurat objects
}
