% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_diff_genes}
\alias{find_diff_genes}
\title{Find differential expressed genes between two groups in each cluster}
\usage{
find_diff_genes(dataset, clusters, comparison, logfc = 0.25)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{clusters}{A string vector - clusters to investigate.}

\item{comparison}{A string vector of length 3 -}

\item{logfc}{A double -}
}
\value{
A Seurat object.
}
\description{
Find differential expressed genes between two groups in each cluster
}
