% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{load_scfile}
\alias{load_scfile}
\title{Load single cell RNA-seq data from directory or file}
\usage{
load_scfile(dir = NULL, gcol = 2, org = "human", cc = TRUE,
  pdx = FALSE, file = NULL, ...)
}
\arguments{
\item{dir}{A string - directory path generated by 10xGenomics pipeline. Default value is \code{NULL}.}

\item{gcol}{An integer - the column index of gene names in the 10xGenomics directory. Default value is \code{2}.}

\item{org}{A string - name of organism. Currently only "human" or "mouse" are accepted. Default value is \code{"human"}.}

\item{cc}{A logcial value - if cell cycle scores are calculated. Default value is \code{TRUE}.}

\item{pdx}{A logcal value - if the sequencing sample is human - mouse mixed. Default value is \code{FALSE}.}

\item{file}{A string - file path of single cell RNA-Seq count matrix. Default value is \code{NULL}. If both \code{dir} and \code{file} are provided. Only the former will be used.}

\item{...}{Additional arguments to be passed to the function \code{\link{CreateSeuratObject}}.}
}
\value{
A Seurat of object.
}
\description{
Load single cell RNA-seq data from directory or file
}
