% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{load_scfile}
\alias{load_scfile}
\title{Load single cell RNA-seq data from directory or file}
\usage{
load_scfile(metadata, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to the function \code{\link{CreateSeuratObject}}.}

\item{meta}{A dataframe - must contain a column named "sample", and a column named "folder" or "file".}
}
\value{
A list of Seurat object
}
\description{
Load single cell RNA-seq data from directory or file
}
