% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_all_cluster_go}
\alias{plot_all_cluster_go}
\title{plot the GO enrichment analysis of all clusters of a dataset}
\usage{
plot_all_cluster_go(markers, org = "human", ...)
}
\arguments{
\item{markers}{A tibble - the dataframe generated by \code{FindAllMarkers} function in \code{Seurat}.}

\item{org}{A string - the organism that the genes belong to. Can take values from \code{c("human","mouse")}.}

\item{...}{Additional arguments to be passed to the function \code{\link{plot_cluster_go}}.}
}
\value{
A plot
}
\description{
plot the GO enrichment analysis of all clusters of a dataset
}
