% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cluster_go}
\alias{plot_cluster_go}
\title{plot the GO enrichment analysis of a cluster}
\usage{
plot_cluster_go(markers, cluster_name, topn = 100, org, ...)
}
\arguments{
\item{markers}{A dataframe - the dataframe generated by \code{FindAllMarkers} function in \code{Seurat}.}

\item{cluster_name}{A string - the name of the single cell cluster.}

\item{topn}{An integer - the number of top gene markers in each cluster for the GO enrichment test.}

\item{org}{A string - the organism that the genes belong to. Can take values from \code{c("human","mouse")}.}

\item{...}{Additional arguments to be passed to the function \code{\link[clusterProfiler]{enrichGO}}.}
}
\value{
A plot.
}
\description{
plot the GO enrichment analysis of a cluster
}
