% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_features}
\alias{plot_features}
\title{Single cell visualization - gene expression levels}
\usage{
plot_features(dataset, genes, split.by = NULL, pt.size = 0.2, ...)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{genes}{A string vector -}

\item{split.by}{A string -}

\item{pt.size}{A double -}

\item{...}{Additional arguments to be passed to the function \code{\link{FeaturePlot}}.}
}
\value{
A plot.
}
\description{
Single cell visualization - gene expression levels
}
