% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_GSEA}
\alias{plot_GSEA}
\title{plot the results of GSEA}
\usage{
plot_GSEA(
  gsea_res,
  pattern = "HALLMARK_",
  p_cutoff = 0.05,
  colors = c("#0570b0", "grey", "#d7301f")
)
}
\arguments{
\item{gsea_res}{A tibble - the results of GSEA generated by \code{test_GSEA}.}

\item{pattern}{A string - the prefix string pattern to be trimmed of pathway names.}

\item{p_cutoff}{A numeric - the p-value cutoff for significance.}

\item{colors}{A string vector - three color values, corresponding to the \code{low}, \code{mid}, and \code{high} of the color gradient.}
}
\value{
A plot.
}
\description{
plot the results of GSEA
}
