% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap of single cell dataset}
\usage{
plot_heatmap(dataset, markers, nfeatures, cluster_pal = c("Paired",
  "Set2", "Set1"), group_colors = c("#92c5de", "#d6604d"))
}
\arguments{
\item{dataset}{A Seurat object.}

\item{markers}{A tibble -}

\item{nfeatures}{An integer -}

\item{cluster_pal}{A string vector -}

\item{group_colors}{A string vector -}
}
\value{
A plot.
}
\description{
Plot the heatmap of single cell dataset
}
