% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap of single cell dataset}
\usage{
plot_heatmap(
  dataset,
  markers,
  sort_var = c("seurat_clusters", "sample"),
  n = 8,
  anno_var,
  anno_colors
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{markers}{A tibble -}

\item{sort_var}{A string vector -}

\item{n}{An integer -}

\item{anno_var}{A string vector -}

\item{anno_colors}{A string vector -}
}
\value{
A plot.
}
\description{
Plot the heatmap of single cell dataset
}
