% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap of single cell dataset}
\usage{
plot_heatmap(dataset, markers, nfeatures = 5, cluster_colors = NULL,
  group_colors = NULL)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{markers}{A tibble -}

\item{nfeatures}{An integer -}

\item{cluster_colors}{A string vector -}

\item{group_colors}{A string vector -}
}
\value{
A plot.
}
\description{
Plot the heatmap of single cell dataset
}
