% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_measure}
\alias{plot_measure}
\title{Box plot/Violin plot of gene expressions or meta measures}
\usage{
plot_measure(dataset, measures, group_by, pal_setup = NULL, show = "combined")
}
\arguments{
\item{dataset}{A Seurat object.}

\item{measures}{A string vector -}

\item{group_by}{A string -}

\item{pal_setup}{A dataframe with 2 columns - the \code{RColorBrewer} palette setup to be used. Default value is \code{NULL}.}

\item{show}{A string -}
}
\value{
A plot.
}
\description{
Box plot/Violin plot of gene expressions or meta measures
}
