% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_measure}
\alias{plot_measure}
\title{Box plot/Violin plot of gene expressions or meta measures}
\usage{
plot_measure(
  dataset,
  features,
  plot_type,
  show = "combined",
  meta = FALSE,
  group_colors = NULL,
  cluster_colors = NULL
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{features}{A string vector -}

\item{plot_type}{A string -}

\item{show}{A string -}

\item{meta}{A logical value -}

\item{group_colors}{A string vector -}

\item{cluster_colors}{A string vector -}
}
\value{
A plot.
}
\description{
Box plot/Violin plot of gene expressions or meta measures
}
