% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_measure}
\alias{plot_measure}
\title{Boxplot of gene program scores}
\usage{
plot_measure(dataset, measure, plot_type, show = "combined",
  group_colors = NULL, cluster_colors = NULL)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{measure}{A string -}

\item{plot_type}{A string -}

\item{show}{A string -}

\item{group_colors}{A string vector -}

\item{cluster_colors}{A string vector -}
}
\value{
A plot.
}
\description{
Boxplot of gene program scores
}
