% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_measure_dim}
\alias{plot_measure_dim}
\title{Single cell visualization - gene expression levels}
\usage{
plot_measure_dim(dataset, measures, split_by = NULL, point.size = 1)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{measures}{A string vector -}

\item{split_by}{A string -}

\item{point.size}{A double -}
}
\value{
A plot.
}
\description{
Single cell visualization - gene expression levels
}
