% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_merge}
\alias{plot_merge}
\title{Single cell visualization - merged and colored by biological sample}
\usage{
plot_merge(dataset, reduction = "umap", group.by = "group",
  colors = c("#92c5de", "#d6604d"), legend.title = "Group",
  labels = levels(dataset$group))
}
\arguments{
\item{dataset}{A Seurat object.}

\item{reduction}{A string -}

\item{group.by}{A string -}

\item{colors}{A string vector -}

\item{legend.title}{A string -}

\item{labels}{A string vector -}
}
\value{
A plot.
}
\description{
Single cell visualization - merged and colored by biological sample
}
