% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_qc}
\alias{plot_qc}
\title{Plot the QC metrics of an individual biological sample}
\usage{
plot_qc(data_list, metrics)
}
\arguments{
\item{data_list}{A list of Seurat objects.}

\item{metrics}{A string - the name of the QC metrics.}
}
\value{
A plot.
}
\description{
Plot the QC metrics of an individual biological sample
}
