% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_scdata}
\alias{plot_scdata}
\title{Single cell visualization - merged and colored by biological sample}
\usage{
plot_scdata(
  dataset,
  color_by = "seurat_clusters",
  split_by = NULL,
  pal_setup = NULL
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{color_by}{A string - by which metadata the colors will be applied. Default value is \code{"seurat_clusters"}.}

\item{split_by}{A string - by which metadata the plot will be split. Default value is \code{NULL}.}

\item{pal_setup}{A dataframe with 2 columns or a string - the \code{RColorBrewer} palette setup to be used. Default value is \code{NULL}.}
}
\value{
A plot.
}
\description{
Single cell visualization - merged and colored by biological sample
}
