% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_scdata}
\alias{plot_scdata}
\title{Single cell visualization - merged and colored by biological sample}
\usage{
plot_scdata(
  dataset,
  color_by = "seurat_clusters",
  split_by = NA,
  pal_setup = "Set2"
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{color_by}{A string - by which metadata the colors will be applied. Default value is \code{"seurat_clusters"}.}

\item{split_by}{A string - by which metadata the plot will be split. Default value is \code{NA}.}

\item{pal_setup}{A dataframe with 2 columns - the \code{RColorBrewer} palette setup to be used, or a string vector of \code{RColorBrewer} palettes, 
or a string vector of colors. Default value is \code{'Set2'} in \code{RColorBrewer}.}
}
\value{
A plot.
}
\description{
Single cell visualization - merged and colored by biological sample
}
