% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_split}
\alias{plot_split}
\title{Single cell visualization - split by biological sample, colored by cluster}
\usage{
plot_split(dataset, color_by = "seurat_clusters", colors = NULL, ...)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{color_by}{A string - by which metadata the colors will be applied. Default value is \code{"seurat_clusters"}.}

\item{colors}{A string vector - the colors used for levels. Default value is \code{NULL}.}
}
\value{
A plot.
}
\description{
Single cell visualization - split by biological sample, colored by cluster
}
