% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_stat}
\alias{plot_stat}
\title{Plot the basic statistics after cell clustering and labeling}
\usage{
plot_stat(
  dataset,
  plot_type,
  cluster_colors = NULL,
  sample_colors = NULL,
  plot_ratio = 1,
  text_size = 10,
  tilt_text = FALSE
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{plot_type}{A string -}

\item{cluster_colors}{A string vector -}

\item{sample_colors}{A string vector -}

\item{plot_ratio}{A double -}

\item{text_size}{A double -}

\item{tilt_text}{A logical -}
}
\value{
A plot.
}
\description{
Plot the basic statistics after cell clustering and labeling
}
