% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{test_GSEA}
\alias{test_GSEA}
\title{GSEA analysis of differential gene expression in each cluster}
\usage{
test_GSEA(diff, clusters = NULL, pathway)
}
\arguments{
\item{diff}{A tibble -}

\item{clusters}{A string vector - clusters to investigate. Default value is \code{NULL}.}

\item{pathway}{A vector list - group of gene lists}
}
\value{
A Seurat object.
}
\description{
GSEA analysis of differential gene expression in each cluster
}
